Add Czech as target language, manual eval and *unsegmented* eval for Czech and German
The discussion with Juan Pino is still ongoing and we understand the goal of keeping the number of evaluation tracks low. However, we fear we wouldn't be able to provide judges for German targets, so we want to have Czech covered. Also, it is easier for us to solicit more participating teams for this target language.

I also propose to move German and the newly-added Czech to *unsegmented* output, because we want to provide human interpretation benchmarks to both, and human interpreters do not obey sentence boundaries.